requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
openai==1.3.7
Pillow==10.1.0
python-decouple==3.8
google-genai==0.4.0
